backend/ms-bible/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0
aiofiles==23.2.1
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 